+++ src_docs/components/example.component.jsx
@@ -33,7 +33,13 @@ export default class ComponentView extends React.PureComponent {
     activeKey: this.state.activeTab,
   });
 
-  getAlignedContent = content => <div>{content}</div>
+  getAlignedContent = content => (
+    <div style={{ 'line-height': '40px' }}>
+      <span style={{ 'vertical-align': 'middle' }}>
+        {content}
+      </span>
+    </div>
+  );
 
   getMenuItems = () => [{
     id: '1',