+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+* Aligned horizontally left aligned content vertically bottom
+* Updated @opuscapita dependencies versions
 
 ## 0.1.1
 * Removed bottom margin